feat, refactor, fix - Added the filtering feature in 1.2 version, changed the gradle dependencies to libs.versions.toml method, fixed some warnings -HomeScreen

+++ app/src/main/java/gr/manolasn/takeaticket/common/utils/Transformations.kt
@@ -47,9 +47,9 @@ object Transformations {
         }
     }
 
-    fun formatBigNumberWithCommas(number: Int): String {
+    fun formatBigNumberWithCommas(number: Long): String {
         val formatter = NumberFormat.getNumberInstance(Locale.US)
-        if (number == 0) return "-"
+        if (number == 0L) return "-"
         return formatter.format(number) + " $"
     }
